Little Refactor
